requirements.txt
asgiref==3.11.1
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cloudinary==1.44.1
coverage==7.13.5
cryptography==46.0.5
dj-database-url==3.1.2
Django==5.2.11
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
idna==3.11
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.1
python-http-client==3.3.7
reportlab==4.4.10
requests==2.32.5
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
webencodings==0.5.1
Werkzeug==3.1.6
whitenoise==6.12.0
